backend/core-api/pyproject.toml
[tool.poetry]
name = "Core-API"
version = "0.1.0"
description = "Core API for the Area project"
authors = ["ItsZcx <joanpaumeridaruiz@gmail.com>"]
readme = "README.md"

package-mode = false

# Add production dependencies (ex: poetry add 'fastapi[standard]')
# Add dev dependencies (ex: poetry add ruff --group=dev)
[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
authlib = "^1.3.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
email-validator = "^2.1.0"
pydantic-settings = "^2.5.2"
google-auth = "^2.23.4"
google-api-python-client = "^2.102.0"
twilio = "^9.3.2"
openai = "^1.51.2"
web3 = "7.4.0"
praw = "^7.8.0"
pytest = "8.3.3"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

####################################
#       Custom ruff settings       #
####################################

[tool.ruff]
# Inherit exclude patterns from ruff.toml
extend-exclude = [".venv", ".env", "alembic"]

# Other settings from ruff.toml
fix = true
line-length = 120
indent-width = 4

target-version = "py310"

[tool.ruff.format]
preview = false
line-ending = "lf"
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 110
skip-magic-trailing-comma = false

[tool.ruff.lint]
fixable = ["ALL"]
ignore = []
preview = false
extend-ignore = []
extend-fixable = []
allowed-confusables = []
explicit-preview-rules = false
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
force-single-line = true
